devtools/client/webconsole/new-console-output/test/components/test_console-api-call.html

<!DOCTYPE HTML>
<html lang="en">
<head>
  <meta charset="utf8">
  <title>Test for ConsoleApiCall component</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript;version=1.8" src="head.js"></script>
  <!-- Any copyright is dedicated to the Public Domain.
     - http://creativecommons.org/publicdomain/zero/1.0/ -->
</head>
<body>
<p>Test for ConsoleApiCall component</p>

<script type="text/javascript;version=1.8">
const { prepareMessage } = require("devtools/client/webconsole/new-console-output/utils/messages");
const { ConsoleApiCall } = require("devtools/client/webconsole/new-console-output/components/message-types/console-api-call");

window.onload = Task.async(function* () {
  yield testConsoleLog();
  yield testConsoleCount();

  SimpleTest.finish()
});

function testConsoleLog() {
  const packet = yield getPacket("console.log('foobar', 'test')", "consoleAPICall");
  const message = prepareMessage(packet);
  const rendered = renderComponent(ConsoleApiCall, {message});

  const messageBody = getMessageBody(rendered);
  // @TODO should output: foobar test
  is(messageBody.textContent, "\"foobar\"\"test\"", "ConsoleApiCall outputs expected text");

  const consoleStringNodes = messageBody.querySelectorAll(".objectBox");
  is(consoleStringNodes.length, 2, "ConsoleApiCall outputs expected HTML structure");
}

function testConsoleCount() {
  for (let i = 0; i < 3; i++) {
    const packet = yield getPacket("console.count('bar')", "consoleAPICall");
    const message = prepareMessage(packet);
    const rendered = renderComponent(ConsoleApiCall, {message: message});
    const messageBody = getMessageBody(rendered);

    const expected = `bar: ${i + 1}`;
    is(messageBody.textContent, expected,
      `console.count has the expected text content: "${expected}"`);
  }

  const packet = yield getPacket("console.count()", "consoleAPICall")
  const message = prepareMessage(packet);
  const rendered = renderComponent(ConsoleApiCall, {message: message});
  const messageBody = getMessageBody(rendered);
  const expected = "<no label>: 1";
  is(messageBody.textContent, expected,
    `console.count without label has the expected text content: "${expected}"`);
}

function getMessageBody(renderedComponent) {
  const queryPath = "div.message.cm-s-mozilla span span.message-flex-body span.message-body.devtools-monospace";
  return renderedComponent.querySelector(queryPath);
}

</script>
</body>
</html>
